Refactor/rename nav components.
